backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0

# Redis
redis[asyncio]==5.0.4

# Milvus
pymilvus==2.3.6

# MinIO
minio==7.2.7

# HTTP
httpx==0.27.0

# Document parsing
unstructured[pdf,docx]==0.13.7
python-magic==0.4.27

# BM25
rank-bm25==0.2.2

# Auth & Security
PyJWT==2.8.0
bcrypt==4.1.3

# Utils
numpy==1.26.4
loguru==0.7.2
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0
